Move TEST_NOCACHE_TIME to /src/constants.ts to fix failed build
