+++ src/domain/events/eventsTable/__tests__/EventsTableRow.test.tsx
@@ -2,7 +2,7 @@ import { MockedResponse } from '@apollo/client/testing';
 import range from 'lodash/range';
 import React from 'react';
 
-import { MAX_PAGE_SIZE } from '../../../../constants';
+import { MAX_PAGE_SIZE, TEST_NOCACHE_TIME } from '../../../../constants';
 import {
   EventFieldsFragment,
   EventsDocument,
@@ -21,7 +21,6 @@ import {
   configure,
   render,
   screen,
-  TEST_NOCACHE_TIME,
   userEvent,
 } from '../../../../utils/testUtils';
 import { EVENT_LIST_INCLUDES } from '../../constants';
