+++ src/utils/getNocacheTime.ts
@@ -1,7 +1,7 @@
 import getUnixTime from 'date-fns/getUnixTime';
 
+import { TEST_NOCACHE_TIME } from '../constants';
 import isTestEnv from './isTestEnv';
-import { TEST_NOCACHE_TIME } from './testUtils';
 
 const getNocacheTime = (): number =>
   isTestEnv ? TEST_NOCACHE_TIME : getUnixTime(new Date());
